Cleaner implem

+++ bin/jarhardlinker.pl
@@ -6,83 +6,95 @@ use Getopt::Long;
 use File::Next;
 use File::Slurp;
 use File::stat;
+use POSIX;
 
 my $nexus="/opt/nexus/sonatype-work/nexus3/blobs/default/content";
-my ($gradle, $maven, $sbt, $debug, $dryRun, $unmanaged);
+my ($all, $gradle, $maven, $sbt, $debug, $dryRun, $printStats, $relink);
 
-GetOptions ("nexus=s"  => \$nexus,   # String
-            "maven"    => \$maven,   # string
-            "sbt"      => \$sbt,     # string
-            "gradle"   => \$gradle,
-            "debug"    => \$debug,
-            "dry-run"  => \$dryRun,
-            "unmanaged" => \$unmanaged
+GetOptions ("nexus=s"      => \$nexus,   # String
+            "maven"        => \$maven,
+            "all"          => \$all,
+            "sbt"          => \$sbt, 
+            "gradle"       => \$gradle,
+            "debug"        => \$debug,
+            "print-stats"  => \$printStats,
+            "dry-run"      => \$dryRun,
+            "relink"       => \$relink
             )   # Numeric
 or die("Error in command line arguments\n");
  
-my $mavenCache = $ENV{"HOME"} . "/.m2/repository/";
-my $sbtCache = $ENV{'HOME'} . '/Library/Caches/Coursier/v1/http/localhost%3A8081/repository/';
-my $gradleCache = $ENV{'HOME'} . '/.gradle/caches/modules-2/files-2.1/';
+my $cachePaths = {
+  maven => $ENV{"HOME"} . "/.m2/repository/",
+  sbt => $ENV{'HOME'} . '/Library/Caches/Coursier/v1/http/localhost%3A8081/repository/',
+  gradle => $ENV{'HOME'} . '/.gradle/caches/modules-2/files-2.1/'
+};
 
-
-unmanaged() if $unmanaged;
-
-my $nexusFiles = buildNexusIndex($nexus);
-
-
-hardLinkFiles("sbt", $sbtCache, sub {
+my $cacheResolvers = {
+  maven => sub {$_[0]},
+  sbt => sub {
     my $relative = shift;
     my ($repo, @paths) = split('/', $relative);
     join('/', @paths);
-}) if $sbt;
-
-hardLinkFiles("mvn", $mavenCache) if $maven;
-hardLinkFiles("gradle", $gradleCache, sub {
+  },
+  gradle => sub {
     my $relative = shift;
     my ($groupId, $artifactId, $version, $sha, $filename) = split('/', $relative);
     $groupId =~ s!\.!/!g;
     join('/', $groupId, $artifactId, $version, $filename);
-} ) if $gradle;
+  }
+};
 
+my @repositories = ();
+
+if($all){
+  @repositories = ("sbt", "maven", "gradle")
+}else{
+  push @repositories, "sbt" if $sbt;
+  push @repositories, "maven" if $maven;
+  push @repositories, "gradle" if $gradle;
+}
+my $nexusFiles = {};
+$nexusFiles = buildNexusIndex($nexus) if $relink && @repositories;
+
+
+foreach my $repository (@repositories) {
+  print $repository, "\n";
+  die "No repository ($repository)\n" unless -d $cachePaths->{$repository};
+
+  printStats($repository, $cachePaths->{$repository}) if $printStats;
+  hardLinkFiles($repository) if $relink;
+}
+
+exit(0);
 
 sub hardLinkFiles(){
-   my $type = shift;
-   my $cachePath = shift;
-   my $resolver = shift
-    || sub {$_[0]};
+   my $repository = shift;
+   my $cachePath = $cachePaths->{$repository};
+   my $resolver = $cacheResolvers->{$repository};
 
-
-   my $stats = {};
    debug("Hard linking $cachePath");
    my $count=0;
    my $sbtIter = File::Next::files({ 'file_filter' => sub { /\.jar$/ } }, $cachePath );
    while ( defined ( my $file = $sbtIter->() ) ) {
        my $relative = substr($file, length($cachePath));
        my $path = $resolver->($relative);
-       $count += harlink($path, $file, $type), $stats;
+       $count += harlink($path, $file, $repository);
    }
-   printStats($type, $count, $cachePath, $stats);
 
 }
 
 sub harlink {
     my $path = shift;
     my $file = shift;
-    my $type = shift;
-    my $stats = shift;
     if($dryRun){
       debug(" 👎 $path") unless $nexusFiles->{$path};
   #    debug(" 👍 $path") if $nexusFiles->{$path};
     }elsif( -e $nexusFiles->{$path}){
       my $st = stat($file);
-#      return 0 if $st->nlink > 1; 
-      $stats->{"newly size red"} += $st->size;
-      $stats->{'newly hardlinked'}++;
+      return if $st->nlink > 1; 
       unlink $file;
       link  $nexusFiles->{$path}, $file;
-      1
     }else{
-     $stats->{'unmanaged'}++;
      debug("[-]  $path"); 
      0
     }
@@ -91,19 +103,19 @@ sub harlink {
 
 
 sub printStats {
-    my $type = shift;
-    my $count = shift;
-    my $cachePath = shift;
-    my $stats=shift;
-    print $type, ': ', $count, "\n";
-    my $sbtIter = File::Next::files({ 'file_filter' => sub { /\.jar$/ } }, $cachePath );
+    my $repository = shift;
+    my $cachePath = $cachePaths->{$repository};
+    my $stats = {};
+    my $sbtIter = File::Next::files({ 'file_filter' => sub { /\.(jar)$/ } }, $cachePath );
     while ( defined ( my $file = $sbtIter->() ) ) {
         my $st = stat($file);
-        next if $st->nlink == 1;
-        $stats->{"size red"} += $st->size;
         $stats->{'nlink '.$st->nlink}++;
-        print $file, "\n" if $st->nlink == 4;
+        next if $st->nlink == 1;
+        $stats->{"size red"} += $st->size/(1024*1024);
     }
+
+    $stats->{"size red"} = floor($stats->{"size red"})." Mo";
+
     foreach my $key  (sort keys %{$stats}) {
         print "\t", $key, ': ', $stats->{$key}, "\n"
     }
@@ -132,30 +144,6 @@ sub buildNexusIndex {
   }
   $index
 }
-
-sub unmanaged {
-  my $stats = {};
-  unmanagedInCache("sbt", $sbtCache, $stats) if $sbt;  
-  foreach my $key  (sort keys %{$stats}) {
-        print "\t", $key, ': ', $stats->{$key}, "\n"
-    }
-  exit
-}
-
-sub unmanagedInCache {
-  my $type = shift;
-  my $cache = shift;
-  my $stats = shift;
-
-  my $iter = jarIterator($cache);
-  while ( defined ( my $file = $iter->() ) ) {
-    my $st = stat($file);
-    next unless $st->nlink == 1;
-    print $file, "\n";
-    $stats->{"unmanaged in $type"}++
-  }
-}
-
 
 sub jarIterator {
   my $cachePath = shift;
